# perf: enable local cache and use `tablepool` to reuse the temporary table