--- a/.luacheckrc
+++ b/.luacheckrc
@@ -1,0 +1,4 @@
+std = "ngx_lua"
+unused_args = false
+redefined = false
+max_line_length = 100