--- a/examples/nginx.conf
+++ b/examples/nginx.conf
@@ -40,11 +40,15 @@ http {
         -- type 'boolean', mark the entrySpan include host/domain
         metadata_buffer:set('includeHostInEntrySpan', false)
 
+        -- set randomseed
         require("skywalking.util").set_randomseed()
+
         require("skywalking.client"):startBackendTimer("http://127.0.0.1:8080")
 
         -- Any time you want to stop reporting metrics, call `destroyBackendTimer`
         -- require("skywalking.client"):destroyBackendTimer()
+
+        skywalking_tracer = require("skywalking.tracer")
     }
 
 
@@ -62,21 +66,21 @@ http {
                 --
                 -- Currently, we can not have the upstream real network address
                 ------------------------------------------------------
-                require("skywalking.tracer"):start("upstream service")
+                skywalking_tracer:start("upstream service")
                 -- If you want correlation custom data to the downstream service
-                -- require("skywalking.tracer"):start("upstream service", {custom = "custom_value"})
+                -- skywalking_tracer:start("upstream service", {custom = "custom_value"})
             }
 
             proxy_pass http://127.0.0.1:8080/tier2/lb;
 
             body_filter_by_lua_block {
                 if ngx.arg[2] then
-                    require("skywalking.tracer"):finish()
+                    skywalking_tracer:finish()
                 end
             }
 
             log_by_lua_block {
-                require("skywalking.tracer"):prepareForReport()
+                skywalking_tracer:prepareForReport()
             }
         }
 
@@ -84,19 +88,19 @@ http {
             default_type text/html;
 
             rewrite_by_lua_block {
-                require("skywalking.tracer"):start("backend service")
+                skywalking_tracer:start("backend service")
             }
 
             proxy_pass http://127.0.0.1:8080/backend;
 
             body_filter_by_lua_block {
                 if ngx.arg[2] then
-                    require("skywalking.tracer"):finish()
+                    skywalking_tracer:finish()
                 end
             }
 
             log_by_lua_block {
-                require("skywalking.tracer"):prepareForReport()
+                skywalking_tracer:prepareForReport()
             }
         }