--- a/examples/nginx.conf
+++ b/examples/nginx.conf
@@ -48,6 +48,10 @@ http {
         -- Any time you want to stop reporting metrics, call `destroyBackendTimer`
         -- require("skywalking.client"):destroyBackendTimer()
 
+        -- If there is a bug of this `tablepool` implementation, we can
+        -- disable it in this way
+        -- require("skywalking.util").disable_tablepool()
+
         skywalking_tracer = require("skywalking.tracer")
     }