--- a/lib/skywalking/client.lua
+++ b/lib/skywalking/client.lua
@@ -15,6 +15,7 @@
 -- limitations under the License.
 --
 
+local ngx = ngx
 local SEGMENT_BATCH_COUNT = 100
 
 local Client = {}
@@ -30,7 +31,6 @@ function Client:startBackendTimer(backend_http_uri)
     local check
 
     local log = ngx.log
-    local DEBUG = ngx.DEBUG
     local ERR = ngx.ERR
 
     check = function(premature)
@@ -101,7 +101,6 @@ end
 -- Ping the backend to update instance heartheat
 function Client:ping(metadata_buffer, backend_http_uri)
     local log = ngx.log
-    local DEBUG = ngx.DEBUG
     local ERR = ngx.ERR
 
     local serviceName = metadata_buffer:get('serviceName')
@@ -136,7 +135,6 @@ end
 -- Send segemnts data to backend
 local function sendSegments(segmentTransform, backend_http_uri)
     local log = ngx.log
-    local DEBUG = ngx.DEBUG
     local ERR = ngx.ERR
 
     local http = require('resty.http')
@@ -167,7 +165,6 @@ end
 function Client:reportTraces(metadata_buffer, backend_http_uri)
     local log = ngx.log
     local DEBUG = ngx.DEBUG
-    local ERR = ngx.ERR
 
     local queue = ngx.shared.tracing_buffer
     local segment = queue:rpop('segment')