--- a/test/e2e/e2e-with-mock-collector/docker/conf.d/nginx.conf
+++ b/test/e2e/e2e-with-mock-collector/docker/conf.d/nginx.conf
@@ -40,6 +40,8 @@ http {
 
         require("skywalking.util").set_randomseed()
         require("skywalking.client"):startBackendTimer("http://${collector}:12800")
+
+        skywalking_tracer = require("skywalking.tracer")
     }
 
     server {
@@ -49,17 +51,17 @@ http {
             default_type text/html;
 
             rewrite_by_lua_block {
-                require("skywalking.tracer"):start("e2e-test-with-mock-collector:upstream_ip:port")
+                skywalking_tracer:start("e2e-test-with-mock-collector:upstream_ip:port")
             }
 
             proxy_pass http://127.0.0.1:8080/tier2/lb;
 
             body_filter_by_lua_block {
-                require("skywalking.tracer"):finish()
+                skywalking_tracer:finish()
             }
 
             log_by_lua_block {
-                require("skywalking.tracer"):prepareForReport()
+                skywalking_tracer:prepareForReport()
             }
         }
 
@@ -67,17 +69,17 @@ http {
             default_type text/html;
 
             rewrite_by_lua_block {
-                require("skywalking.tracer"):start("e2e-test-with-mock-collector:upstream_ip2:port2")
+                skywalking_tracer:start("e2e-test-with-mock-collector:upstream_ip2:port2")
             }
 
             proxy_pass http://127.0.0.1:8080/backend;
 
             body_filter_by_lua_block {
-                require("skywalking.tracer"):finish()
+                skywalking_tracer:finish()
             }
 
             log_by_lua_block {
-                require("skywalking.tracer"):prepareForReport()
+                skywalking_tracer:prepareForReport()
             }
         }