--- a/test/e2e/e2e-with-mock-collector/docker/conf.d/nginx.conf
+++ b/test/e2e/e2e-with-mock-collector/docker/conf.d/nginx.conf
@@ -41,6 +41,10 @@ http {
         require("skywalking.util").set_randomseed()
         require("skywalking.client"):startBackendTimer("http://${collector}:12800")
 
+        -- If there is a bug of this `tablepool` implementation, we can
+        -- disable it in this way
+        -- require("skywalking.util").disable_tablepool()
+
         skywalking_tracer = require("skywalking.tracer")
     }