this PR depends on: https://github.com/apache/skywalking-nginx-lua/pull/50