@membphis 2021-02-21
The test result obtained when the skywalking service is not started. 

```shell
# this PR
➜  skywalking-nginx-lua git:(perf) ✗ wrk http://127.0.0.1:8080/ingress
Running 10s test @ http://127.0.0.1:8080/ingress
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   738.99us  415.02us   5.29ms   92.24%
    Req/Sec     7.30k     1.77k    8.27k    88.12%
  146615 requests in 10.10s, 36.77MB read
Requests/sec:  14516.75
Transfer/sec:      3.64MB

# master branch
➜  skywalking-nginx-lua-upstream git:(master) ✗ wrk http://127.0.0.1:8080/ingress
Running 10s test @ http://127.0.0.1:8080/ingress
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   801.90us  459.00us   4.60ms   91.62%
    Req/Sec     6.74k     1.71k    7.68k    87.13%
  135492 requests in 10.10s, 33.98MB read
Requests/sec:  13416.06
Transfer/sec:      3.36MB
```