@membphis 2021-02-21
The test result obtained when the skywalking service is running. 

```shell
# current PR
➜  skywalking-nginx-lua git:(perf) ✗ wrk http://127.0.0.1:8080/ingress
Running 10s test @ http://127.0.0.1:8080/ingress
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   697.70us  248.50us   2.57ms   90.55%
    Req/Sec     7.26k     0.92k    8.39k    51.00%
  144390 requests in 10.00s, 36.21MB read
Requests/sec:  14436.94
Transfer/sec:      3.62MB

# master branch
➜  skywalking-nginx-lua-upstream git:(master) ✗ wrk http://127.0.0.1:8080/ingress
Running 10s test @ http://127.0.0.1:8080/ingress
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   738.83us  275.64us   3.41ms   89.22%
    Req/Sec     6.86k     0.93k   11.76k    82.59%
  137225 requests in 10.10s, 34.41MB read
Requests/sec:  13586.21
Transfer/sec:      3.41MB
```